+++ axelor-stock/src/main/java/com/axelor/apps/stock/module/StockModule.java
@@ -45,6 +45,8 @@ import com.axelor.apps.stock.service.PartnerProductQualityRatingService;
 import com.axelor.apps.stock.service.PartnerProductQualityRatingServiceImpl;
 import com.axelor.apps.stock.service.PartnerStockSettingsService;
 import com.axelor.apps.stock.service.PartnerStockSettingsServiceImpl;
+import com.axelor.apps.stock.service.StockCorrectionService;
+import com.axelor.apps.stock.service.StockCorrectionServiceImpl;
 import com.axelor.apps.stock.service.StockLocationLineService;
 import com.axelor.apps.stock.service.StockLocationLineServiceImpl;
 import com.axelor.apps.stock.service.StockLocationService;
@@ -94,5 +96,6 @@ public class StockModule extends AxelorModule {
     bind(PickingStockMovePrintService.class).to(PickingStockMovePrintServiceimpl.class);
     bind(ConformityCertificatePrintService.class).to(ConformityCertificatePrintServiceImpl.class);
     bind(StockLocationLineRepository.class).to(StockLocationLineStockRepository.class);
+    bind(StockCorrectionService.class).to(StockCorrectionServiceImpl.class);
   }
 }